+++ example/test/integration/rails_fixtures_override_test.rb
@@ -6,15 +6,11 @@ class FixtureFreeTestData < ActiveSupport::TestCase
   fixtures :boops # why not
 
   setup do
-    TestData.load
+    TestData.uses_test_data
   end
 
   def test_has_no_fixture_boops
     assert_equal 15, Boop.count
-  end
-
-  teardown do
-    TestData.rollback
   end
 end
 
@@ -22,7 +18,7 @@ class FixturesUsingTest < ActiveSupport::TestCase
   fixtures :boops
 
   setup do
-    TestData.load_rails_fixtures(self)
+    TestData.uses_rails_fixtures(self)
   end
 
   def test_has_fixture_boops
@@ -35,8 +31,8 @@ class FixturesUsingTest < ActiveSupport::TestCase
   end
 
   def test_even_explicitly_loading_test_data_will_truncate_and_then_load_fixtures
-    TestData.load
-    TestData.load_rails_fixtures(self)
+    TestData.uses_test_data
+    TestData.uses_rails_fixtures(self)
 
     assert_equal 2, Boop.count
   end
@@ -50,14 +46,10 @@ class FixturesUsingTest < ActiveSupport::TestCase
 
     assert_equal Boop.find(boop.id).created_at.to_date, a_year_ago
 
-    # Now after rollback
-    TestData.rollback(:after_load_rails_fixtures)
+    # Now trigger a rollback to the fixtures point
+    TestData.uses_rails_fixtures(self)
 
     assert_equal Boop.find(boop.id).created_at.to_date, original_created_on
-  end
-
-  teardown do
-    TestData.rollback(:after_load_rails_fixtures)
   end
 end
 
@@ -66,35 +58,29 @@ class SomeFixturesAndSomeTestDataInOneClassTest < ActiveSupport::TestCase
   fixtures :all
 
   def test_fixtures_work
-    TestData.load_rails_fixtures(self)
+    TestData.uses_rails_fixtures(self)
 
     assert_equal Date.civil(2020, 1, 1), boops(:boop_1).updated_at.to_date
     assert_equal "Levi", pants(:pant_1).brand
-
-    TestData.rollback(:after_load_rails_fixtures)
   end
 
   def test_that_rewinds_to_test_data
-    TestData.load
+    TestData.uses_test_data
 
     assert_equal 15, Boop.count
-
-    TestData.rollback
   end
 
   def test_that_rewinds_to_the_very_start
-    TestData.rollback(:before_data_load)
+    TestData.uninitialize
 
     assert_equal 0, Boop.count
   end
 
   def test_fixtures_get_reloaded_because_cache_is_cleared
-    TestData.load_rails_fixtures(self)
+    TestData.uses_rails_fixtures(self)
 
     assert_equal Date.civil(2019, 1, 1), boops(:boop_2).updated_at.to_date
     assert_equal "Wrangler", pants(:pant_2).brand
-
-    TestData.rollback(:after_load_rails_fixtures)
   end
 end
 
@@ -102,11 +88,7 @@ class PantsFixturesTest < ActiveSupport::TestCase
   fixtures :pants
 
   setup do
-    TestData.load_rails_fixtures(self)
-  end
-
-  teardown do
-    TestData.rollback(:after_load_rails_fixtures)
+    TestData.uses_rails_fixtures(self)
   end
 
   def test_has_fixture_pants
@@ -121,8 +103,8 @@ end
 class FixtureTestPassingTheWrongThingTest < ActiveSupport::TestCase
   def test_doing_it_wrong
     error = assert_raises(TestData::Error) do
-      TestData.load_rails_fixtures(ActiveRecord::Base)
+      TestData.uses_rails_fixtures(ActiveRecord::Base)
     end
-    assert_match "'TestData.load_rails_fixtures' must be passed a test instance that has had ActiveRecord::TestFixtures mixed-in (e.g. `TestData.load_rails_fixtures(self)` in an ActiveSupport::TestCase `setup` block), but the provided argument does not respond to 'setup_fixtures'", error.message
+    assert_match "'TestData.uses_rails_fixtures(self)' must be passed a test instance that has had ActiveRecord::TestFixtures mixed-in (e.g. `TestData.uses_rails_fixtures(self)` in an ActiveSupport::TestCase `setup` block), but the provided argument does not respond to 'setup_fixtures'", error.message
   end
 end