+++ example/test/integration/transaction_committing_boops_test.rb
@@ -1,13 +1,11 @@
 require "test_helper"
 
-TestData.config.use_transactional_data_loader = false
-
 class TransactionCommittingTestCase < ActiveSupport::TestCase
   self.use_transactional_tests = false
 
   setup do
     Noncommittal.stop!
-    TestData.load
+    TestData.insert_test_data_dump
   end
 
   teardown do
@@ -25,12 +23,5 @@ class TransactionCommittingBoopsTest < TransactionCommittingTestCase
 
   def test_finds_the_boops_via_another_process
     assert_equal 15, `RAILS_ENV=test bin/rails runner "puts Boop.count"`.chomp.to_i
-  end
-
-  def test_cant_have_it_both_ways
-    error = assert_raise(TestData::Error) do
-      TestData.config.use_transactional_data_loader = true
-    end
-    assert_match "There is already a non-transactional data loader", error.message
   end
 end